Validation fixes
To try and fix all the errors with the teacher-created code we downloaded the project files anew and just added our own code as a test solution. This did not fix the problem, but as the problem is with the teacher-created code it does not affect our code or the project in general.